+++ ddtrace/contrib/langchain/patch.py
@@ -27,7 +27,7 @@ from ddtrace.contrib.trace_utils import unwrap
 from ddtrace.contrib.trace_utils import with_traced_module
 from ddtrace.contrib.trace_utils import wrap
 from ddtrace.internal.agent import get_stats_url
-from ddtrace.internal.llmobs.llm_integrations import LangChainIntegration
+from ddtrace.internal.llmobs.integrations import LangChainIntegration
 from ddtrace.internal.logger import get_logger
 from ddtrace.internal.utils import ArgumentError
 from ddtrace.internal.utils import get_argument_value
