+++ ddtrace/internal/llmobs/integrations/base.py
