+++ tests/contrib/openai/conftest.py
@@ -100,7 +100,7 @@ class FilterOrg(TraceFilter):
 
 @pytest.fixture(scope="session")
 def mock_metrics():
-    patcher = mock.patch("ddtrace.internal.llmobs.llm_integrations.base.get_dogstatsd_client")
+    patcher = mock.patch("ddtrace.internal.llmobs.integrations.base.get_dogstatsd_client")
     DogStatsdMock = patcher.start()
     m = mock.MagicMock()
     DogStatsdMock.return_value = m
@@ -114,7 +114,7 @@ def mock_logs(scope="session"):
     Note that this fixture must be ordered BEFORE mock_tracer as it needs to patch the log writer
     before it is instantiated.
     """
-    patcher = mock.patch("ddtrace.internal.llmobs.llm_integrations.base.V2LogWriter")
+    patcher = mock.patch("ddtrace.internal.llmobs.integrations.base.V2LogWriter")
     V2LogWriterMock = patcher.start()
     m = mock.MagicMock()
     V2LogWriterMock.return_value = m
@@ -124,7 +124,7 @@ def mock_logs(scope="session"):
 
 @pytest.fixture
 def mock_llmobs_writer(scope="session"):
-    patcher = mock.patch("ddtrace.internal.llmobs.llm_integrations.base.LLMObsWriter")
+    patcher = mock.patch("ddtrace.internal.llmobs.integrations.base.LLMObsWriter")
     LLMObsWriterMock = patcher.start()
     m = mock.MagicMock()
     LLMObsWriterMock.return_value = m
